requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
flake8==7.2.0
frozenlist==1.6.0
greenlet==3.2.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.5.0
propcache==0.3.1
pyasn1==0.4.8
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
yarl==1.20.0
